requirements.txt
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.5.1
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
django-recaptcha==1.0
django-registration==1.0
nose==1.3.3
numpy==1.8.1
pyparsing==2.0.2
python-dateutil==2.2
six==1.7.3
tornado==4.0
wsgiref==0.1.2
